Do not try to attempt to automatically create meta course link
This basically reverts the feature implemented by Matt Gibson in the pull
request #5. I found it a bit buggy but more importantly, conceptually wrong.
This module should do one thing and do it well. Any attempt to do something on
behalf of the course teacher may lead to unpredictable results (e.g. the fact
that the module deleted the meta course link on its removal).