Do not update all users' grades on triggered events
When a single user was graded or enrolled into the referenced course,
the event triggered update of all users. This had significant impact on
the performance.

The patch allows to update the grade for a single user only. In which
case, the fetch timestamp is not recorded.

This should help with the issue #14