Do not open in another new window if we just got a new one
This is covering a case of the module viewed by a teacher with "Open in
a new window" enabled. It led to a new window opened with the instance
description (as if the redirect was disabled), where the button opened
the course in yet another new window.

To avoid that, the link from the course main page makes view.php aware
of the new window was just opened so it is ok to stay in it.